+++ CHANGELOG.md
@@ -65,6 +65,9 @@ For instructions on upgrading to newer versions, visit
 * \#1610 When versioning paranoid documents and max version is set, hard
   delete old versions from the embedded relation.
 
+* \#1609 Allow connection retry during cursor iteration as well as all other
+  operations.
+
 * \#1608 Guard against no method errors when passing ids in nested attributes
   and the documents do not exist.